parser: fix bug with string beginning with parens.
If a string started with a parenthesis, then it was getting
parsed as a parenthesis token and not a string token. This
fix is a temporary work around until some type information
can be added to the token. This fixes strings like "(TEST)"
but still fails with single character strings "(" and ")".