configure.ac: Add subdir-objects to the automake flags
The unit tests depend on files in another directory, and automake 1.14+
will complain when subdir-objects is not set and such a setup is used.
Combined with -Werror, this leads to autoreconf -i failing with a recent
automake.

Since subdir-objects have no ill side-effects, this patch adds that to
the AM_INIT_AUTOMAKE list of options.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>