Use a convenience library to help the tests
Instead of compiling Edgar twice, once for the REPL, once for the tests,
create a convenience (static) library called libedgar, and link both the
REPL and the test suite to it. This also makes it possible to build
edgar in a subdirectory (like during distcheck) and still have
subdir-objects enabled.

The downside is that this adds a dependency on ar, but that should be
available even on mingw32, I believe.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>